Update Vue components and add FontAwesome icon
This commit has made changes to two Vue components (ArrayValue.vue and UsageValue.vue) to make UI interactions more seamless. A 'v-if' conditional rendering is added in ArrayValue.vue to fix the display issue when the array is collapsed. UsageValue.vue has been updated to streamline padding and table display style. Finally, a new icon 'faMemory' has been imported and added to the FontAwesomeIcon library in fontawsome_icon.js for use in the interface.