update"Fix possible undefined reference in BackTrace component"
In the BackTrace component, we added an optional chaining operator for the "args" property of the "row" object inside a conditional check. This ensures that if "args" is undefined, we won't attempt to access its "type" property, thus avoiding a potential JavaScript error of "Cannot read property 'type' of undefined". This change enhances the robustness of our component against unexpected data structures.