Refactor DataRow and add new ThrowableValue component
The DataRow component has been refactored to simplify the control flow structure for different message types. The code related to throwable message type has been moved to a newly created ThrowableValue component. This makes DataRow cleaner and easier to understand. Moreover, it commences the use of dedicated components for different message types, enhancing the modularity and reusability of the code.

In addition, the CodeViewer component declaration now includes a component name, aligning it with the Vue's style guide recommendation for clearer debugging.