Refactor DataRow and RowExtensions components
This commit simplifies the DataRow component by removing redundant wrapping div elements around the RowExtensions component. This change significantly improves the readability of the component's structure. This commit also updates the RowExtensions component to accept a "languageVersion" slot. The slot is used for displaying language version alongside the DataRow. This change makes DataRow and RowExtensions components more maintainable and clean.