Add keyboard shortcuts for reversing list and switching containers
Keyboard shortcuts were added to the scrollable view and App components. Now, the 'r' key can be used, along with either the 'meta' or 'ctrl' key, to reverse a list in the scrollable view. The 'meta' or 'ctrl' key paired with the '1', '2', or '3' numerical keys can be used to switch between the 'logContainer', 'throwableContainer' or 'shiftContainer' respectively in the App component. These changes improve user experience by offering a quicker and more efficient way to interact with the interface.