Implement dimmed effect when folding code
The update includes the renaming of the emit event 'handleCollapsed' to 'handleFold' in CodeFoldingTail.vue for more semantic clarity. A new utility 'dimmed' has been added to add a dimmed effect when the code is currently in a process of folding. This was done by modifying the 'CodeFoldingTail.vue', 'StdClassValue.vue', and 'ArrayValue.vue' files. The logic of handling code folding has been adjusted to add the dimming effect using useDimmed utility. Also added CSS for dimmed effect in 'global.scss'.