Update display logic and fix command syntax
The display logic method for ScrollableView was removed and replaced by directly computing the display id in the v-for directive. This was done to streamline the code and remove unnecessary complexity.

Also, there was a typo in the Composer command syntax on the 'HelloDocument' Vue component. The correct syntax should be 'composer global require', not 'composer require global'.