"Add timestamp feature and refactor RowExtensions component"
This commit adds a visual timestamp feature on RowExtensions.vue component by leveraging vue's computed property for this purpose. Now, the current date-time is displayed next to a blue circle which gradually fades out over 30 seconds, providing a nice visual effect.

On the refactoring end, an adjustment was made to determine whether data is local or not, changing from "isLocalData" to "props.isLocalData". Also, the component changed from options API to composition API using <script setup> and the separate 'saved' reactive property was created. This provides more clarity and efficiency in our code execution. Animation for the fade-out feature was added on SCSS block.