Increase visibility and readability of buttons and text
The button styling and text within components were slightly difficult to see with previous sizing and color setting. The font size and color saturation of buttons and text-tags have been increased from 80% to 90% to enhance readability. Additionally, saving to the local state has been corrected in the RowExtensions.vue file to avoid confusion and errors.