Remove DOCTYPE declarations from SVG icons
The DOCTYPE declarations have been removed from a set of SVG image icon files. This has been done to simplify the SVG content, as these declarations are not necessary for SVG files to function correctly. This simplification can help to improve the load time and performance of the websites/web apps that use these icons.