Add ping util, data listener implementation, and new icons
Added a ping utility to create a local network call for checking server availability. Implemented a data listener switch to check whether the application is data listening or not. This provides control to the user for pausing/starting data listening when needed. New icons have been introduced for the play and pause functionalities. Also, updated some styling for better usability in dark mode.