Add createdAt property and date handling in RowExtensions and DataRow
Added a new optional property 'createdAt' in RowExtensions. This modification will handle both the local and server time depending on the 'isLocalData' property. For DataRow, 'createdAt' is now passed in case of local data is being used. This change will ensure capturing the correct datetime when saving data to locales storage, making it easier to track changes in data over time.