Update UI for RowExtensions component
Modified the color behavior for a UI element in the RowExtensions component. This change was implemented to improve the visibility and user experience while using the 'noticeCircle' function. Now, when 'noticeCircle' is false, the element will be a neutral gray color instead of invisible. When 'noticeCircle' is true, the color will be a noticeable purple with an animation effect.