Add feedback balloon for user actions
This commit introduces a feedback balloon which pops up to inform users about the outcome of their actions. The "Save to local storage", "Copy as JSON" and "Copy as Assoc" functions in the "DataRow.vue" file now emit an event to display a feedback balloon whenever these actions are completed. This provides a more intuitive user experience, ensuring users are aware that their required action has been performed successfully.

In addition to this, changes were made to the "RowExtensions.vue" file to remove the "saved" status being displayed, as this is now handled by the feedback balloon. The "FontAwesomeIcon.js" and "ScrollableView.vue" files were also updated to accommodate the feedback balloon functionality. Finally, the "FeedbackBalloon.vue" component was created for the feedback balloon feature.